create temperature watcher

+++ src/jvmMain/kotlin/org/tnt/tempwatchertracker/FileTemperatureLogger.kt
@@ -3,7 +3,7 @@ package org.tnt.tempwatchertracker
 import org.tnt.tempwatchertracker.InMemoryTemperatureLogger.TemperatureLogger
 
 class FileTemperatureLogger : TemperatureLogger {
-    override fun logTemperature(temperature: Float) {
+    override fun logTemperature(temperature: Double) {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 }
